Restructure repo as a proposal for future managability
