+++ src/main/scala/org/apache/spark/sql/registrar/functions_datasketches_kll.scala
@@ -17,25 +17,14 @@
 
 package org.apache.spark.sql
 
-
-import org.apache.spark.sql.aggregate.{KllDoublesSketchAgg, KllDoublesMergeAgg}
-import org.apache.spark.sql.aggregate.{ThetaSketchBuild, ThetaUnion}
-import org.apache.spark.sql.catalyst.expressions.{Expression, Literal}
-import org.apache.spark.sql.catalyst.expressions.aggregate.AggregateFunction
 import org.apache.spark.sql.catalyst.expressions.Literal
-import org.apache.spark.sql.expressions._
 import org.apache.spark.sql.functions.lit
 import org.apache.spark.sql.types.{ArrayType, BooleanType, DoubleType}
 
-// this class defines and maps all the variants of each function invocation, analagous
-// to the functions object in org.apache.spark.sql.functions
-object functions_ds {
+import org.apache.spark.sql.aggregate.{KllDoublesMergeAgg, KllDoublesSketchAgg}
+import org.apache.spark.sql.expressions.{KllGetMin, KllGetMax, KllGetPmfCdf}
 
-  private def withExpr(expr: => Expression): Column = Column(expr)
-
-  private def withAggregateFunction(func: AggregateFunction): Column = {
-    Column(func.toAggregateExpression())
-  }
+object functions_datasketches_kll extends DatasketchesScalaFunctionBase {
 
   // get min
   def kll_get_min(expr: Column): Column = withExpr {
@@ -162,47 +151,5 @@ object functions_ds {
 
   def kll_get_cdf(columnName: String, splitPoints: Array[Double]): Column = {
     kll_get_cdf(Column(columnName), splitPoints)
-  }
-
-  // Theta
-
-  def theta_sketch_build(column: Column, lgk: Int): Column = withAggregateFunction {
-    new ThetaSketchBuild(column.expr, lgk)
-  }
-
-  def theta_sketch_build(columnName: String, lgk: Int): Column = {
-    theta_sketch_build(Column(columnName), lgk)
-  }
-
-  def theta_sketch_build(column: Column): Column = withAggregateFunction {
-    new ThetaSketchBuild(column.expr)
-  }
-
-  def theta_sketch_build(columnName: String): Column = {
-    theta_sketch_build(Column(columnName))
-  }
-
-  def theta_union(column: Column, lgk: Int): Column = withAggregateFunction {
-    new ThetaUnion(column.expr, lit(lgk).expr)
-  }
-
-  def theta_union(columnName: String, lgk: Int): Column = withAggregateFunction {
-    new ThetaUnion(Column(columnName).expr, lit(lgk).expr)
-  }
-
-  def theta_union(column: Column): Column = withAggregateFunction {
-    new ThetaUnion(column.expr)
-  }
-
-  def theta_union(columnName: String): Column = withAggregateFunction {
-    new ThetaUnion(Column(columnName).expr)
-  }
-
-  def theta_sketch_get_estimate(column: Column): Column = withExpr {
-    new ThetaSketchGetEstimate(column.expr)
-  }
-
-  def theta_sketch_get_estimate(columnName: String): Column = {
-    theta_sketch_get_estimate(Column(columnName))
   }
 }
