+++ src/main/scala/org/apache/spark/sql/registrar/functions_datasketches_theta.scala
@@ -1,0 +1,65 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql
+
+import org.apache.spark.sql.functions.lit
+
+import org.apache.spark.sql.aggregate.{ThetaSketchBuild, ThetaUnion}
+import org.apache.spark.sql.expressions.ThetaSketchGetEstimate
+
+object functions_datasketches_theta extends DatasketchesScalaFunctionBase {
+  def theta_sketch_build(column: Column, lgk: Int): Column = withAggregateFunction {
+    new ThetaSketchBuild(column.expr, lgk)
+  }
+
+  def theta_sketch_build(columnName: String, lgk: Int): Column = {
+    theta_sketch_build(Column(columnName), lgk)
+  }
+
+  def theta_sketch_build(column: Column): Column = withAggregateFunction {
+    new ThetaSketchBuild(column.expr)
+  }
+
+  def theta_sketch_build(columnName: String): Column = {
+    theta_sketch_build(Column(columnName))
+  }
+
+  def theta_union(column: Column, lgk: Int): Column = withAggregateFunction {
+    new ThetaUnion(column.expr, lit(lgk).expr)
+  }
+
+  def theta_union(columnName: String, lgk: Int): Column = withAggregateFunction {
+    new ThetaUnion(Column(columnName).expr, lit(lgk).expr)
+  }
+
+  def theta_union(column: Column): Column = withAggregateFunction {
+    new ThetaUnion(column.expr)
+  }
+
+  def theta_union(columnName: String): Column = withAggregateFunction {
+    new ThetaUnion(Column(columnName).expr)
+  }
+
+  def theta_sketch_get_estimate(column: Column): Column = withExpr {
+    new ThetaSketchGetEstimate(column.expr)
+  }
+
+  def theta_sketch_get_estimate(columnName: String): Column = {
+    theta_sketch_get_estimate(Column(columnName))
+  }
+}
