+++ src/main/scala/org/apache/spark/sql/theta/aggregate/ThetaUnion.scala
